api/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.104.1
gpt4all==2.0.2
pydantic==2.5.2
PyYAML==6.0.1
requests==2.31.0
